fix(type): fix .vue file type error
