+++ src/components/Form/src/types/index.ts
@@ -105,7 +105,6 @@ export type ComponentType =
   | 'RangePicker'
   | 'WeekPicker'
   | 'TimePicker'
-  | 'ImageUpload'
   | 'Switch'
   | 'StrengthMeter'
   | 'Upload'
